+++ packages/messenger-widget/src/components/SignIn/SignIn.css
@@ -69,9 +69,68 @@
     visibility: hidden;
 }
 
-@media only screen and (max-width: 767px) {
+
+/* =================== Mobile Responsive CSS =================== */
+
+@media only screen and (max-width: 800px) {
+
+    .rainbow-connect-btn {
+        padding: 0.5rem 0.5rem 0rem 0rem !important;
+    }
+
     .home-image-container {
-        height: 0% !important;
+        height: 35% !important;
+        width: 100% !important;
+    }
+
+    .signin-data-container{
+        height: 65%;
+        width: 100%;
+    }
+
+    .details{
+        line-height: 17px !important;
+    }
+
+    .sign-in-logo {
+        width: 50% !important;
+    }
+
+    .wal-not-connected {
+        padding: 0.4rem 1rem 0.4rem 1rem !important;
+        font-size: 0.7rem !important;
+    }
+
+    .content-data-container {
+        padding: 0rem 2rem 0rem 2rem !important;
+    }
+
+    .small-spinner {
+        width: 1.5rem !important;
+    }
+
+    .signin-btn {
+        font-size: 13px !important;
+        padding: 0.5rem 1.2rem 0.7rem 1.2rem !important;
+    }
+
+}
+
+@media only screen and (max-width: 300px) {
+    
+    .details{
+        font-size: 10px !important;
+        line-height: 14px !important;
+    }
+
+    .encrypted-details{
+        font-size: 10px !important;
+        line-height: 14px !important;
+        margin-bottom: 0.5rem !important;
+    }
+
+    .para-div{
+        margin-top: 0.5rem !important;
     }
 }
 
@@ -98,12 +157,6 @@
     }
 }
 
-@container signinContainer (max-width: 640px) {
-    .rainbow-connect-btn {
-        display: none !important;
-    }
-}
-
 @container signinContainer (max-height: 325px) and (height > 0px) {
     .rainbow-connect-btn {
         display: none !important;
@@ -112,8 +165,8 @@
 
 @container signinContainer (max-width: 735px) {
     .signin-btn {
-        font-size: 1rem !important;
-        padding: 0.5rem !important;
+        font-size: 13px !important;
+        padding: 0.5rem 1.2rem 0.7rem 1.2rem !important;
     }
     .loading-btn {
         height: 18px !important;
@@ -131,5 +184,3 @@
         padding: 0rem 8rem 0rem 8rem !important;
     }
 }
-
-/* ==================== Mobile Responsive CSS ==================== */
